[ts] PATHFINDER FOR OPTION VALIDATION WAT (.... 1625)
  • this is such a ridiculous tiny corner, but there was something
immensely satisifying about getting this working.
  • (as simple as it is, this is the first thing that really starts to
  feel like rudimentary AI)

(4 in 88.52)

+++ lib/skylab/cov-tree/api/actions/sub-tree.rb
@@ -55,10 +55,14 @@ module Skylab::CovTree
 
     def process_pair in_pn, out_pn
       -> do  # #result-block
-        if ! ( dn = out_pn.dirname ).exist?
-          mkdir_p dn, verbose: true, noop: @is_dry_run
+        if in_pn.exist?
+          if ! ( dn = out_pn.dirname ).exist?
+            mkdir_p dn, verbose: true, noop: @is_dry_run
+          end
+          cp in_pn.to_s, out_pn.to_s, verbose: true, noop: @is_dry_run
+        else
+          @err.puts "(DID NOT EXIST, SKIPPING: #{ in_pn })"
         end
-        cp in_pn.to_s, out_pn.to_s, verbose: true, noop: @is_dry_run
         true
       end.call
     end
